% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCongruentBranches.R
\name{GetCongruentBranches}
\alias{GetCongruentBranches}
\title{gets branch length estimates from congruent branches}
\usage{
GetCongruentBranches(
  species_tree_name,
  gene_trees_direc,
  zero_sample_overrule,
  loci_lengths,
  output_tree_name
)
}
\description{
gets branch length estimates from congruent branches
}
